Ammending Network Environments - still not 100% working but progress
