+++ Data Generation/target_factory.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+from typing import Tuple
+
+@dataclass
+class Target:
+    target_id: int
+    target_label: str
+    position: Tuple[float, float]
